pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "pgvector>=0.4.2",
    "psycopg>=3.3.2",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
